+++ src/OpenTelemetry.PersistentStorage.FileSystem/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.0.0-beta.2 (Unreleased)
 
+* `FileBlobProvider` will now use the path provided during initialization as is
+* for storing blobs, without adding additional hash of current user and process.
+([#1110](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1110))
+
 * Going forward the NuGet package will be
   [`OpenTelemetry.PersistentStorage.FileSystem`](https://www.nuget.org/packages/OpenTelemetry.Extensions.FileSystem).
   Older versions will remain at